+++ game/cards/dm05/spells.go
@@ -2,6 +2,7 @@ package dm05
 
 import (
 	"duel-masters/game/civ"
+	"duel-masters/game/cnd"
 	"duel-masters/game/fx"
 	"duel-masters/game/match"
 	"fmt"
@@ -19,14 +20,12 @@ func EnchantedSoil(c *match.Card) {
 
 		if match.AmICasted(card, ctx) {
 
-			cards := match.Search(card.Player, ctx.Match, card.Player, match.GRAVEYARD, "Select 2 cards from your graveyard and put it in your manazone", 0, 2, true)
-
-			for _, card := range cards {
-
+			fx.SelectFilter(card.Player, ctx.Match, card.Player, match.GRAVEYARD, "Enchanted Soil: Select 2 creatures from your graveyard and put it in your manazone", 0, 2, true, func(x *match.Card) bool {
+				return x.HasCondition(cnd.Creature)
+			}).Map(func(x *match.Card) {
 				card.Player.MoveCard(card.ID, match.GRAVEYARD, match.MANAZONE)
 				ctx.Match.Chat("Server", fmt.Sprintf("%s was moved to %s's manazone", card.Name, card.Player.Username()))
-
-			}
+			})
 
 		}
 	})